Move ddb client to store dir and middleware to handler dir
