+++ go_api/store/user.go
@@ -4,8 +4,9 @@ import (
 	"context"
 	"fmt"
 	"strconv"
-	"the-drink-almanac-api/client"
+
 	"the-drink-almanac-api/model"
+	"the-drink-almanac-api/store/client"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue"
